Redeploy the same artifact version if the content changed
Store the releases in similar fashion to how Berkshelf stores the
cookbooks, i.e. extract the artifact to a release directory obtained
by appending checksum to the version number. Compare manifests against
the extracted version, so it is possible to re-deploy the same version
number if the archive content changed.